==> MovieLens/03-Collaborative-Filtering.ipynb <==
# Collaborative Filtering - Intuition

import numpy as np
import pandas as pd

ratings = pd.read_csv("data/ratings.csv")

ratings.head()

## Explicit Prediction Rating

items = pd.read_csv("data/items.csv")
users = pd.read_csv("data/users.csv")
ratings = pd.read_csv("data/ratings.csv")

items.head()

Small Data

| movie_id | title                 |
| -------: | :-------------------- |
|      1   | Toy Story (1995)      |
|     71   | Lion King, The (1994) |
|     95   | Aladdin (1992)        |
|     50   | Star Wars (1972)      |
|    176   | Aliens (1986)         |
|     82   | Jurassic Park (1993)  |



sample_movie_id = [1, 71, 95, 50, 176, 82]
sample_user_id = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15]

sample_items = items[items.movie_id.isin(sample_movie_id)]
sample_users = users[users.user_id.isin(sample_user_id)]

sample_items.head()

#items.query("Action == 1 and SciFi == 1")
#items.query("Children == 1 and Animation == 1")

sample_users

rating_collect = ratings.user_id.isin(sample_user_id) & ratings.movie_id.isin(sample_movie_id)

sample_ratings = ratings[rating_collect]

sample_ratings.head()

len(sample_ratings.user_id.unique())

pd.crosstab(index=sample_ratings.user_id, columns=sample_ratings.movie_id, values=sample_ratings.rating, aggfunc="mean")

import keras

#
