==> MovieLens/06-Model-Baseline.ipynb <==
# Model - Baseline

import sys
sys.path.append("../")

import numpy as np
import pandas as pd
from IPython.display import SVG
import matplotlib.pyplot as plt
import seaborn as sns

from reco.preprocess import encode_user_item, random_split, user_split

%matplotlib inline

## Prepare Data

df_ratings = pd.read_csv("data/ratings.csv")

df_ratings.head()

DATA = encode_user_item(df_ratings, "user_id", "movie_id", "rating", "unix_timestamp")

DATA.head()

# Data Splitting
#train, test = user_split(DATA, "TIMESTAMP", "USER", [0.75, 0.25])
train, test = random_split(DATA, [0.75, 0.25])

train.shape, test.shape

## Build & Train Model 

- Baseline (Rating): We will use each user's **mean rating** as the baseline prediction.
- Baseline (Ranking): We will use the **top-k popular items** as the baseline ranking


def average_rating_model(train):
    users_ratings = train.groupby(["USER"])["RATING"].mean()
    users_ratings = users_ratings.reset_index()
    users_ratings.rename(columns = {'RATING': 'RATING_PRED'}, inplace = True)

    return users_ratings

def popular_item_model(train):
    item_counts = (train.groupby("ITEM")
                   .count()
                   .reset_index()
                   .sort_values(ascending = False, by = "USER"))
    item_counts = item_counts[["ITEM", "USER"]]
    item_counts.columns = ['ITEM', 'RATING_PRED']
        
    return item_counts

### Generate Prediction: Rating & Ranking

predictions_ratings = average_rating_model(train)
predictions_ratings.head()

predictions_ranking = popular_item_model(train)
predictions_ranking.head()

## Evaluate Rating & Ranking 

rating_evaluate_df = pd.merge(test, predictions_ratings, on=['USER'], how='inner')
rating_evaluate_df.head()

from sklearn.metrics import mean_absolute_error, mean_squared_error

mean_absolute_error(rating_evaluate_df.RATING, rating_evaluate_df.RATING_PRED)

### For Ranking

from reco.evaluate import user_item_crossjoin, filter_by, get_top_k_items
from reco.evaluate import precision_at_k, recall_at_k, ndcg_at_k

def recommend_topk(data, train, k=5):
    
    """
    Params:
        data (pandas.DataFrame): DataFrame of entire rating data
        train (pandas.DataFrame): DataFrame of train rating data
        k (int): number of items for each user

    Returns:
        pd.DataFrame: DataFrame of top k items for each user, sorted by `col_user` and `rank`
    
    """
    
    # Create the crossjoin for user-item
    user_item = user_item_crossjoin(data)
    
    # Get ranking prediction for each user-item,
    all_predictions = pd.merge(user_item, predictions_ranking, on="ITEM", how="left")
    
    # Handle Missing Values
    all_predictions.fillna(0, inplace=True)
    
    # Filter already seen items
    all_predictions_unseen = filter_by(all_predictions, train, ["USER", "ITEM"])
    
    
    recommend_topk_df = get_top_k_items(all_predictions_unseen, "USER", "RATING_PRED", k=5)
    
    return recommend_topk_df


ranking_topk = recommend_topk(DATA, train, 10)

ranking_topk.head()

eval_precision = precision_at_k(test, ranking_topk, k=10)
eval_recall = recall_at_k(test, ranking_topk, k=10)
eval_ndcg = ndcg_at_k(test, ranking_topk, k=10)

print("NDCG@K:\t%f" % eval_ndcg,
      "Precision@K:\t%f" % eval_precision,
      "Recall@K:\t%f" % eval_recall, sep='\n')
